Creating new covid_key app to manage all code related functions
