+++ covid_key/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class CovidKeyConfig(AppConfig):
+    name = 'covid_key'
