+++ profiles/tests.py
@@ -92,6 +92,17 @@ class HomePageView(TestCase):
         self.assertContains(response, "Welcome to the COVID Alert Portal")
 
 
+class AuthenticatedView(AdminUserTestCase):
+    def test_loginpage(self):
+        #  Get the login page
+        response = self.client.get(reverse("login"))
+        self.assertEqual(response.status_code, 200)
+        self.assertContains(response, "<h1>Log in</h1>")
+        #  Test logging in
+        response = self.client.post("/en/login/", self.credentials, follow=True)
+        self.assertTrue(response.context["user"].is_active)
+
+
 class RestrictedPageViews(TestCase):
     #  These should redirect us
     def test_code(self):
@@ -159,52 +170,6 @@ class DjangoAdminPanelView(AdminUserTestCase):
             HTTP_USER_AGENT="test-browser",
         )
         self.assertEqual(response.status_code, 403)
-
-
-class AuthenticatedView(AdminUserTestCase):
-    def test_loginpage(self):
-        #  Get the login page
-        response = self.client.get(reverse("login"))
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, "<h1>Log in</h1>")
-        #  Test logging in
-        response = self.client.post("/en/login/", self.credentials, follow=True)
-        self.assertTrue(response.context["user"].is_active)
-
-    def test_key(self):
-        """
-        Login and then see the key page and one generated code
-        """
-        self.client.login(username="test@test.com", password="testpassword")
-        response = self.client.get(reverse("key"))
-        self.assertEqual(response.status_code, 302)
-
-        self.login_2fa()
-
-        response = self.client.get(reverse("key"))
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, "<h1>Give patient this key</h1>")
-        self.assertContains(
-            response, "<code>{}</code>".format(response.context["code"])
-        )
-
-    def test_key_instructions(self):
-        """
-        Login and then see the key instructions view and one generated code
-        """
-        self.client.login(username="test@test.com", password="testpassword")
-        response = self.client.get(reverse("key_instructions"))
-        self.assertEqual(response.status_code, 302)
-
-        self.login_2fa()
-
-        response = self.client.get(reverse("key_instructions"))
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, "<h1>Give patient this key</h1>")
-        self.assertContains(response, "<h2>Instructions for patient</h2>")
-        self.assertContains(
-            response, "<code>{}</code>".format(response.context["code"])
-        )
 
 
 class i18nTestView(TestCase):
