+++ profiles/views.py
@@ -1,9 +1,6 @@
-import requests
-import logging
-from datetime import timedelta
 
 from django.conf import settings
-from django.shortcuts import render, redirect
+from django.shortcuts import redirect
 from django.contrib.sites.shortcuts import get_current_site
 from django.utils.translation import gettext as _
 from django.views.generic import (
@@ -14,13 +11,10 @@ from django.views.generic import (
     DetailView,
 )
 from django.contrib import messages
-from django.contrib.auth.decorators import login_required
 from django.urls import reverse_lazy
 from django_otp import DEVICE_ID_SESSION_KEY
-from django_otp.decorators import otp_required
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.db.models.expressions import RawSQL
-from django.utils import timezone
 
 
 from invitations.models import Invitation
@@ -39,8 +33,6 @@ from .forms import (
     Resend2FACodeForm,
 )
 from .utils import get_site_name
-
-logger = logging.getLogger(__name__)
 
 
 class SignUpView(FormView):
@@ -214,44 +206,3 @@ class UserDeleteView(Is2FAMixin, ProvinceAdminDeleteMixin, DeleteView):
     model = HealthcareUser
     context_object_name = "profile_user"
     success_url = reverse_lazy("profiles")
-
-
-@login_required
-@otp_required
-def code(request):
-    token = settings.API_AUTHORIZATION
-    diagnosis_code = "0000000000"
-    if token:
-        try:
-            r = requests.post(
-                settings.API_ENDPOINT, headers={"Authorization": f"Bearer {token}"}
-            )
-            r.raise_for_status()  # If we don't get a valid response, throw an exception
-            # Make sure the code has a length of 10, cheap sanity check
-            if len(r.text.strip()) == 10:
-                diagnosis_code = r.text
-            else:
-                logger.error(
-                    f"The key API returned a key with the wrong format : {r.text}"
-                )
-        except requests.exceptions.HTTPError as err:
-            logging.exception(
-                f"Received {r.status_code} with message {err.response.text}"
-            )
-        except requests.exceptions.RequestException as err:
-            logging.exception(f"Something went wrong {err}")
-
-    # Split up the code with a space in the middle so it looks like this: 123 456 789
-    diagnosis_code = (
-        f"{diagnosis_code[0:3]} {diagnosis_code[3:6]} {diagnosis_code[6:10]}"
-    )
-
-    expiry = timezone.now() + timedelta(days=1)
-
-    template_name = "key_instructions" if "/key-instructions" in request.path else "key"
-
-    return render(
-        request,
-        "profiles/{}.html".format(template_name),
-        {"code": diagnosis_code, "expiry": expiry},
-    )
